+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Setting ha_discovery_enabled to False had no effect
 * Docker image based on python:3.11-slim
 * Force refresh by charging station only if charging value has changed
+* MQTT connection error logging
 
 ## 0.4.5